Clean up of the project
Removed all unrelated parts that do not belong to the primary task of the tool. Also fixed handling of certain Java statement processing parts and split the `processBodyMembers` method, which became pretty large.

The Java polyfill library path can now be set via options too. With this setup it is now possible to keep the target code in a completely separate (project) folder.